Lucene.Class

Lucene language

class Lucene expr query | query -> expr, expr -> query where

The finally tagless Lucene class. This admits multiple interpreters, with one (lazy ByteStrings) provided by this library, in the Lucene.Query module.

For simplicity, the type signatures in the examples below monomorphise the functions to use LuceneQuery (and therefore LuceneExpr, due to the functional dependency).

Methods

int :: Int -> expr TInt

An int expression.

Note that sometimes you may use the Num instance for LuceneExpr TInt, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:5
query :: LuceneQuery
query = "foo" =: int 5

true :: expr TBool

A true expression.

Example:

-- foo:true
query :: LuceneQuery
query = "foo" =: true

false :: expr TBool

A false expression.

Example:

-- foo:false
query :: LuceneQuery
query = "foo" =: false

word :: Text -> expr TWord

A single word. Must not contain any spaces, wildcard characters ('?' and '*'), or tildes ('~'), though this is not enforced by the type system.

Note that sometimes you may use the IsString instance for LuceneExpr TWord, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example:

-- foo:bar
query :: LuceneQuery
query = "foo" =: word "bar"

Or, with OverloadedStrings:

-- foo:bar
query :: LuceneQuery
query = "foo" =: ("bar" :: LuceneExpr TWord)

wild :: Text -> expr TWild

A single word that may contain wildcard characters ('?' and '*'), although it must not begin with a '*'. Must also not contain any spaces or tildes ('~'), though this is not enforced by the type system.

Example:

-- foo:b?r
query :: LuceneQuery
query = "foo" =: wild "b?r"

regex :: Text -> expr TRegex

A regular expression, whose syntax is described by http://lucene.apache.org/core/5_5_0/core/org/apache/lucene/util/automaton/RegExp.html?is-external=true.

Note that the leading and trailing '/' must be omitted. The regex innards are not type checked in any way.

-- foo:/[mb]oat/
query :: LuceneQuery
query = "foo" =: regex "[mb]oat"

phrase :: [expr TWord] -> expr TPhrase

A phrase, composed of multiple (non-fuzzy) words, none of which may contain wildcard characters. Both of these properties are enforced by the type system, as long as the words themselves adhere to the word contract. The list should not be empty.

Note that sometimes you may use the IsList instance for LuceneExpr TPhrase, but usually an explicit type signature will be required (at the interpretation site or earlier).

Example (with OverloadedStrings):

-- foo:"bar baz"
query :: LuceneQuery
query = "foo" =: phrase ["bar", "baz"] -- ok

-- foo:"bar b?z" (an invalid Lucene query)
query :: LuceneQuery
query = "foo" =: phrase ["bar", wild "b?z"] -- type error

-- foo:"bar b?z" (an invalid Lucene query)
query :: LuceneQuery
query = "foo" =: phrase ["bar", "b?z"] -- breaks word contract

Or, with OverloadedLists and OverloadedStrings:

-- foo:"bar baz"
query :: LuceneQuery
query = "foo" =: (["bar", "baz"] :: LuceneExpr TPhrase)

fuzz :: FuzzableType a => expr a -> Int -> expr (TFuzzed a)

The '~' operator, which fuzzes its argument (either a word or phrase) by a numeric amount.

This will have one of the following two types:

fuzz :: LuceneExpr TWord   -> Int -> LuceneExpr TFuzzyWord   -- Int must be 0, 1, or 2
fuzz :: LuceneExpr TPhrase -> Int -> LuceneExpr TFuzzyPhrase -- Int must be positive

Example:

-- foo:bar~1
query :: LuceneQuery
query = "foo" =: fuzz (word "bar") 1

-- foo:"bar baz qux"~10
query :: LuceneQuery
query = "foo" =: fuzz (phrase ["bar", "baz", "qux"]) 10

to :: PrimType a => Boundary (expr a) -> Boundary (expr a) -> expr TRange infixr 5

A range expression.

This will have one of the following two types:

to :: Boundary (LuceneExpr TWord) -> Boundary (LuceneExpr TWord) -> LuceneExpr TRange
to :: Boundary (LuceneExpr TInt)  -> Boundary (LuceneExpr TInt)  -> LuceneExpr TRange

Example:

-- foo:[5 TO 10}
query :: LuceneQuery
query = "foo" =: to (incl (int 5)) (excl (int 10))

boost :: BoostableType a => expr a -> Float -> expr (TBoosted a)

A boost expression.

This will have one of the following two types:

boost :: LuceneExpr TWord   -> Float -> LuceneExpr TBoostedWord
boost :: LuceneExpr TPhrase -> Float -> LuceneExpr TBoostedPhrase

Example:

-- foo:bar^3.5
query :: LuceneQuery
query = "foo" =: boost (word "bar") 3.5

-- foo:"bar baz"^3.5
query :: LuceneQuery
query = "foo" =: boost (phrase ["bar", "baz"]) 3.5

(=:) :: Text -> expr a -> query infix 4

A field query.

Example:

-- foo:bar
query :: LuceneQuery
query = "foo" =: word "bar"

(&&:) :: query -> query -> query infixr 3

An AND query.

Example:

-- foo:bar AND baz:qux
query :: LuceneQuery
query = "foo" =: word "bar"
    &&: "baz" =: word "qux"

(||:) :: query -> query -> query infixr 2

An OR query.

Example:

-- foo:bar OR baz:qux
query :: LuceneQuery
query = "foo" =: word "bar"
    ||: "baz" =: word "qux"

(-:) :: query -> query -> query infixr 1

A NOT or '-' query.

Example:

-- foo:bar -baz:qux
query :: LuceneQuery
query = "foo" =: word "bar"
     -: "baz" =: word "qux"

Derived combinators

(~:) :: (Lucene expr query, FuzzableType a) => expr a -> Int -> expr (TFuzzed a) infixr 5

An infix version of fuzz.

fuzzy :: Lucene expr query => expr TWord -> expr TFuzzyWord

Short-hand for fuzzing a word by 2. This is the default behavior of a Lucene '~' operator without an integer added.

fuzzy e = fuzz e 2

Example:

-- foo:bar~
query :: LuceneQuery
query = "foo" =: fuzzy "bar"

(^:) :: (Lucene expr query, BoostableType a) => expr a -> Float -> expr (TBoosted a) infixr 5

An infix version of boost.

gt :: (Lucene expr query, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than range query.

gt e = excl e `to` star

Example:

-- foo:>5
-- foo:{5 TO *]
query :: LuceneQuery
query = "foo" =: gt (int 5)

gte :: (Lucene expr query, PrimType a) => expr a -> expr TRange

Short-hand for a greater-than-or-equal-to range query.

gte e = incl e `to` star

Example:

-- foo:>=5
-- foo:[5 TO *]
query :: LuceneQuery
query = "foo" =: gt (int 5)

lt :: (Lucene expr query, PrimType a) => expr a -> expr TRange

Short-hand for a less-than range query.

 lt e = star `to` excl e

Example:

-- foo:<5
-- foo:[* TO 5}
query :: LuceneQuery
query = "foo" =: lt (int 5)

lte :: (Lucene expr query, PrimType a) => expr a -> expr TRange

Short-hand for a less-than-or-equal-to range query.

lte e = star `to` incl e

Example:

-- foo:<=5
-- foo:[* TO 5]
query :: LuceneQuery
query = "foo" =: lte (int 5)

Range expression helpers

data Boundary a

An inclusive or exclusive expression for use in a range query, built with either incl, excl, or star.

The constructors are exported for use in interpreters.

Constructors

Inclusive a

Exclusive a

Star

incl :: Lucene expr query => expr a -> Boundary (expr a)

Mark an expression as inclusive, for use in a range query.

excl :: Lucene expr query => expr a -> Boundary (expr a)

Mark an expression as exclusive, for use in a range query.

star :: Lucene expr query => Boundary (expr a)