Lucene.Expr.Typed

Expression type

data LuceneExpr :: LuceneType -> *

A type-tagged Lucene expression.

Expression construction

int :: Int -> LuceneExpr TInt

An int expression.

Note that sometimes you may use the Num instance for LuceneExpr TInt, however, due to the interaction between type classes and GADTs, an explicit type signature will usually be required.

Example:

-- foo:5
query :: LuceneQuery
query = "foo" =: int 5

Here's an example where only one integer has to be explicitly "wrapped" with int, for better or for worse:

-- foo:[5 TO 6]
query :: LuceneQuery
query = "foo" =: range (incl (int 5)) (incl 6)

true :: LuceneExpr TBool

A true expression.

Example:

-- foo:true
query :: LuceneQuery
query = "foo" =: true

false :: LuceneExpr TBool

A false expression.

Example:

-- foo:false
query :: LuceneQuery
query = "foo" =: false

word :: Text -> LuceneExpr TWord

A single word. Must not contain any spaces, wildcard characters ('?' and '*'), or tildes ('~'), though this is not enforced by the type system.

Note that sometimes you may use the IsString instance for LuceneExpr TWord, however, due to the interaction between type classes and GADTs, an explicit type signature will usually be required.

Example:

-- foo:bar
query :: LuceneQuery
query = "foo" =: word "bar"

Or, with OverloadedStrings:

-- foo:bar
query :: LuceneQuery
query = "foo" =: ("bar" :: LuceneExpr TWord)

wild :: Text -> LuceneExpr TWild

A single word that may contain wildcard characters ('?' and '*'), although it must not begin with a '*'. Must also not contain any spaces or tildes ('~'), though this is not enforced by the type system.

Example:

-- foo:b?r
query :: LuceneQuery
query = "foo" =: wild "b?r"

regex :: Text -> LuceneExpr TRegex

A regular expression, whose syntax is described by http://lucene.apache.org/core/5_5_0/core/org/apache/lucene/util/automaton/RegExp.html?is-external=true.

Note that the leading and trailing '/' must be omitted. The regex innards are not type checked in any way.

-- foo:/[mb]oat/
query :: LuceneQuery
query = "foo" =: regex "[mb]oat"

phrase :: [LuceneExpr TWord] -> LuceneExpr TPhrase

A phrase, composed of multiple (non-fuzzy) words, none of which may contain wildcard characters. Both of these properties are enforced by the type system, as long as the words themselves adhere to the word contract. The list should not be empty.

Note that sometimes you may use the IsList instance for LuceneExpr TPhrase, however, due to the interaction between type classes and GADTs, an explicit type signature will usually be required.

Example (with OverloadedStrings):

-- foo:"bar baz"
query :: LuceneQuery
query = "foo" =: phrase ["bar", "baz"] -- ok

-- foo:"bar b?z" (an invalid Lucene query)
query :: LuceneQuery
query = "foo" =: phrase ["bar", wild "b?z"] -- type error

-- foo:"bar b?z" (an invalid Lucene query)
query :: LuceneQuery
query = "foo" =: phrase ["bar", "b?z"] -- breaks word contract

Or, with OverloadedLists and OverloadedStrings:

-- foo:"bar baz"
query :: LuceneQuery
query = "foo" =: (["bar", "baz"] :: LuceneExpr TPhrase)

fuzz :: FuzzableType a => LuceneExpr a -> Int -> LuceneExpr (TFuzzed a)

A '~' operator, which fuzzes its argument (either a word or phrase) by a numeric amount.

This will have one of the following two types:

fuzz :: LuceneExpr TWord   -> Int -> LuceneExpr TFuzzyWord   -- Int must be 0, 1, or 2
fuzz :: LuceneExpr TPhrase -> Int -> LuceneExpr TFuzzyPhrase -- Int must be positive

Example:

-- foo:bar~1
query :: LuceneQuery
query = "foo" =: fuzz (word "bar") 1

-- foo:"bar baz qux"~10
query :: LuceneQuery
query = "foo" =: fuzz (phrase ["bar", "baz", "qux"]) 10

(~:) :: FuzzableType a => LuceneExpr a -> Int -> LuceneExpr (TFuzzed a) infixr 5

An infix version of fuzz.

fuzzy :: LuceneExpr TWord -> LuceneExpr TFuzzyWord

Short-hand for fuzzing a word by 2. This is the default behavior of a Lucene '~' suffix without an integer added.

Example:

-- foo:bar~
query :: LuceneQuery
query = "foo" =: fuzzy "bar" -- equivalent to: word "bar" ~: 2

to :: PrimType a => Boundary (LuceneExpr a) -> Boundary (LuceneExpr a) -> LuceneExpr TRange infixr 5

A range expression.

This will have one of the following two types:

to :: Boundary (LuceneExpr TWord) -> Boundary (LuceneExpr TWord) -> LuceneExpr TRange
to :: Boundary (LuceneExpr TInt)  -> Boundary (LuceneExpr TInt)  -> LuceneExpr TRange

Example:

-- foo:[5 TO 10}
query :: LuceneQuery
query = "foo" =: to (incl (int 5)) (excl (int 10))

boost :: BoostableType a => LuceneExpr a -> Float -> LuceneExpr (TBoosted a)

A boost expression.

This will have one of the following two types:

boost :: LuceneExpr TWord   -> Float -> LuceneExpr TBoostedWord
boost :: LuceneExpr TPhrase -> Float -> LuceneExpr TBoostedPhrase

Example:

-- foo:bar^3.5
query :: LuceneQuery
query = "foo" =: boost (word "bar") 3.5

-- foo:"bar baz"^3.5
query :: LuceneQuery
query = "foo" =: boost (phrase ["bar", "baz"]) 3.5

(^:) :: BoostableType a => LuceneExpr a -> Float -> LuceneExpr (TBoosted a) infixr 5

An infix version of boost.

data Boundary a

An inclusive or exclusive expression for use in a range query, built with either incl or excl.

incl :: LuceneExpr a -> Boundary (LuceneExpr a)

Mark an expression as inclusive, for use in a range query.

excl :: LuceneExpr a -> Boundary (LuceneExpr a)

Mark an expression as exclusive, for use in a range query.

Expression compilation

compileLuceneExpr :: LuceneExpr a -> Builder

Compile a Lucene expression to a Builder. Because a Lucene expression is correct by construction, this function is total.